client-3-4/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=1.10,<3.0

# Scheduling
apscheduler>=3.10,<4.0

# Date/time handling
python-dateutil>=2.8.0

# LED Matrix display (rgbmatrix for standard Raspberry Pi)
# Note: This typically needs to be built from source
# See: https://github.com/hzeller/rpi-rgb-led-matrix/tree/master/bindings/python
# If not already installed, you may need to follow the installation instructions there

# ADS-B decoding
pyModeS>=2.10

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Streamlit web interface for configuration
streamlit>=1.28.0